Migrate from Example to Test
