+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/sacloud/simplemq-api-go
 
-go 1.24.3
+go 1.24
 
 tool github.com/ogen-go/ogen/cmd/ogen
